+++ datatig/models/field.py
@@ -8,12 +8,14 @@ class FieldConfigModel(ABC):
         self._title = None
         self._description: str = ""
         self._extra_config = {}
+        self._required: bool = False
 
     def load(self, config: dict) -> None:
         self._id = config.get("id")
         self._title = config.get("title", self._id)
         self._description = config.get("description", "")
         self._key = config.get("key")
+        self._required = config.get("required", False)
         self._load_extra_config(config)
 
     def _load_extra_config(self, config: dict) -> None:
@@ -24,6 +26,7 @@ class FieldConfigModel(ABC):
         self._title = data["title"]
         self._description = data["description"]
         self._key = data["key"]
+        self._required = data["required"]
 
     def get_id(self) -> str:
         return self._id
@@ -40,6 +43,9 @@ class FieldConfigModel(ABC):
     def get_extra_config(self) -> dict:
         return self._extra_config
 
+    def get_required(self) -> bool:
+        return self._required
+
     def get_value_object(self, record, data):
         pass
 
@@ -48,6 +54,9 @@ class FieldConfigModel(ABC):
 
     def get_frictionless_csv_resource_specifications(self) -> list:
         return []
+
+    def get_json_schema(self) -> dict:
+        return {}
 
 
 class FieldValueModel(ABC):