+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `unique_items` option to `list-dictionaries` and `list-string` field types https://github.com/DataTig/DataTig/issues/39
 * `min_length` and `max_length` options to `string` field type https://github.com/DataTig/DataTig/issues/40
 * `string_min_length` and `string_max_length` options to `list-strings` field type https://github.com/DataTig/DataTig/issues/37
-
+* `required` option to all field types https://github.com/DataTig/DataTig/issues/42
+* ... and `required` column to `type_field` table in SQLite database
 
 ### Changed